pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "adis-secrets"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[project.optional-dependencies]
aws = ["boto3"]
gcp = ["google-cloud-secret-manager"]
